feat(#138): move bytecode classes to upper level
