+++ src/main/java/org/eolang/jeo/representation/generation/BytecodeClass.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.jeo.representation.asm.generation;
+package org.eolang.jeo.representation.generation;
 
 import com.jcabi.xml.XML;
 import java.io.PrintWriter;
